# src/main/java/ru/mishelby/walletapi/model/DepositOperationRequest.java
package ru.mishelby.walletapi.model;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;

public record DepositOperationRequest(
        @NotNull(message = "Amount must not be null")
        @Positive(message = "Amount must be greater than zero")
        BigDecimal amount
) {
}